+++ R/Matched.R
@@ -35,28 +35,35 @@
 #' `MatchedSummarizedExperiment` function providing the `SummarizedExperiment`
 #' with the `query` parameter.
 #'
-#' A `Matched` or `MatchedSummarizedExperiment` object can be subsetted in the
-#' following ways:
-#'
 #' - `[` subset the object selecting `query` object elements to keep with
 #'   parameter `i`. The resulting object will contain all the matches
 #'   for the selected query elements. The `target` object will by default be
 #'   returned as-is.
 #'
+#' - `addMatches`: add new matches to an existing object. Parameters
+#'   `queryValue` and `targetValue` allow to define which element(s) in
+#'   `query` and `target` should be considered matching. If `isIndex = TRUE`,
+#'   both `queryValue` and `targetValue` are considered to be integer indices
+#'   identifying the matching elements in `query` and `target`, respectively.
+#'   Alternatively (with `isIndex = FALSE`) `queryValue` and `targetValue` can
+#'   be elements in columns `queryColumn` or `targetColumn` which can be used to
+#'   identify the matching elements. Note that in this case
+#'   **only the first** matching pair is added. Parameter `score` allows to
+#'   provide the score for the match. It has to be a `data.frame` with the same
+#'   columns than `@matches` (except columns `"query_idx"` and `"target_idx"`).
+#'   See examples below for more information.
+#'
+#' - `filterMatches`: keeps or removes matches corresponding to certain indexes
+#'   or values of `query` and `target`. If `queryValue` and `targetValue` are
+#'   provided, matches for these value pairs are kept or removed. Parameter
+#'   `index` allows to filter matches providing their index in the `@matches`
+#'   slot. Note that `filterMatches` removes only matches from the `@matches`
+#'   slot from the `Matched` object but thus not alter the `query` or `target`
+#'   in the object. See examples below for more information.
+#'
 #' - `pruneTarget` *cleans* the object by removing non-matched
 #'   **target** elements.
-#'   
-#' - `filterMatches` keeps only the matches corresponding to certain indexes or
-#'   values of `query` and `target`.
-#'   
-#' It is possible to add new matches to a `Matched` or 
-#' `MatchedSummarizedExperiment` object via `addMatches`. With parameters 
-#' `queryValues` and `targetValues` it is possible to specify values of `query` 
-#' and `target` (if they are 1-dimensional) or one of their columns (specified 
-#' with parameters `queryColname` or `targetColname`) and these values are used 
-#' to establish new matches. By using parameter `isIndex` it is possible to 
-#' interpret `queryValues` and `targetValues` as vector of indexes of the 
-#' elements of `query` and `target` to be matched together.
+#'
 #'
 #' @section Extracting data:
 #'
@@ -110,10 +117,16 @@
 #'
 #' @param i `integer` or `logical` defining the `query` elements to keep.
 #'
+#' @param index for `filterMatches`: indices of the matches to keep (if
+#'  `keep = TRUE`) or to drop if (`keep = FALSE`).
+#'
+#' @param isIndex for `addMatches`: specifies if `queryValue` and
+#' `targetValue` are expected to be vectors of indexes.
+#'
 #' @param j for `[`: ignored.
 #'
-#' @param idxs for `filterMatches`: indexes of the matches to keep (if 
-#'  `keep = TRUE`) or to drop if (`keep = FALSE`).
+#' @param keep for `filterMatches`: `logical`. If `keep = TRUE` the matches are
+#' kept, if `keep = FALSE` they are removed.
 #'
 #' @param matches `data.frame` with columns `"query_idx"` (`integer`),
 #'   `"target_idx"` (`integer`) and `"score"` (`numeric`) representing the n:m
@@ -123,40 +136,38 @@
 #'
 #' @param object a `Matched` object.
 #'
+#' @param score for `addMatches`: `data.frame` (same number of rows than
+#'     `queryValue`) specifying the scores for the matches to add.
+#'
 #' @param target object with the elements against which `query` has been
 #'   matched.
 #'
+#' @param targetColname if `query` is 2-dimensional: column of `target` against
+#'   which elements of `targetValue` are compared.
+#'
+#' @param targetValue for `filterMatches`: vector of values to search for in
+#'   `target` (if `target` is 1-dimensional) or in column `targetColname` of
+#'   `target` (if `target` is 2-dimensional). For `addMatches`: either an
+#'   index in `target` or value in column `targetColumn` of `target` defining
+#'   (together with `queryValue`) the pair of query and target elements for
+#'   which a match should be manually added. Lengths of `queryValue` and
+#'   `targetValue` have to match.
+#'
 #' @param query object with the query elements.
 #'
-#' @param x `Matched` object.
-#' 
-#' @param queryValues vector of values to search for in `query` (if `query` 
-#' is 1-dimensional) or in column `queryColname` of `query` (if `query` is  
-#' 2-dimensional). This parameter can be used in `filterMatches` 
-#' and when it is used also the parameter `targetValues` must be passed. 
-#' `filterMatches` will return a `Matched` object with only the matches whose 
-#' `query` and  `target` values are equal respectivly to `targetValues[i]` and 
-#' `queryValues[i]` for some i.
-#' 
-#' @param targetValues vector of values to search for in `target` (if `target` 
-#' is 1-dimensional) or in column `targetColname` of `target` (if `target` is 
-#' 2-dimensional). 
-#' 
-#' @param queryColname if `query` is 2-dimensional: column of `query` against 
-#' which elements of `queryValues` are compared.
-#' 
-#' @param targetColname if `query` is 2-dimensional: column of `target` against 
-#' which elements of `targetValues` are compared.
+#' @param queryColname if `query` is 2-dimensional: column of `query` against
+#'   which elements of `queryValue` are compared.
 #'
-#' @param keep for `filterMatches`: `logical`. If `keep = TRUE` the matches are 
-#' kept, if `keep = FALSE` they are removed.
-#' 
-#' @param isIndex for `addMatches`: specifies if `queryValues` and 
-#' `targetValues` are expected to be vectors of indexes.
-#' 
-#' @param scores for `addMatches`: data.frame specifying the scores for the 
-#' matches to add.
-#' 
+#' @param queryValue for `filterMatches`: vector of values to search for in
+#'   `query` (if `query` is 1-dimensional) or in column `queryColname` of
+#'   `query` (if `query` is 2-dimensional). For `addMatches`: either an index
+#'   in `query` or value in column `queryColumn` of `query` defining (together
+#'   with `targetValue`) the pair of query and target elements for which a
+#'   match should be manually added. Lengths of `queryValue` and
+#'   `targetValue` have to match.
+#'
+#' @param x `Matched` object.
+#'
 #' @param ... additional parameters.
 #'
 #' @return See individual method description above for details.
@@ -178,7 +189,7 @@
 #' ## with target rows 2,3,4 and query row 5 with target row 5.
 #' mo <- Matched(
 #'     q1, t1, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
-#'                                 target_idx = c(2L, 2L, 3L, 4L, 5L),
+#'                                  target_idx = c(2L, 2L, 3L, 4L, 5L),
 #'                                  score = seq(0.5, 0.9, by = 0.1)))
 #' mo
 #'
@@ -235,6 +246,7 @@
 #' mo_sub <- pruneTarget(mo_sub)
 #' nrow(target(mo_sub))
 #'
+#' ########
 #' ## Creating a Matched object with a `data.frame` for `query` and a `vector`
 #' ## for `target`. The matches are specified in the same way as the example
 #' ## before.
@@ -258,7 +270,6 @@
 #'
 #' ## Note that the columns of the obtained `DataFrame` are the same as the
 #' ## corresponding vectors obtained with `$`
-#'
 #' res$col1
 #' res$target
 #'
@@ -278,34 +289,55 @@
 #'
 #' ## Reducing the target elements to only those that match at least one query row
 #' mo_sub <- pruneTarget(mo_sub)
-#' nrow(target(mo_sub))
+#' length(target(mo_sub))
 #'
-#' ## Keeping only certain matches through their indexes 
-#' idxs <- c(1, 3, 5)
-#' mosub <- filterMatches(mo, idxs = idxs)
-#' mosub@matches
-#' mo@matches[-idxs, ]
-#' 
-#' ## Keeping matches corresponding to certain values of query and target
-#' queryValues <- c(q1[mo@matches[idxs, "query_idx"], "col1"], -1, - 2)
-#' targetValues <- c(t2[mo@matches[idxs, "target_idx"]], -2, -1)
-#' mosub <- filterMatches(mo, queryValues = queryValues, 
-#'                      targetValues = targetValues, queryColname = "col1")
-#' mosub@matches
+#' ########
+#' ## Filtering `Matched` with `filterMatches`
 #'
-#' ## Removing matches using values of query and target
-#' mosub <- filterMatches(mo, keep = FALSE, queryValues = queryValues, 
-#'                      targetValues = targetValues, queryColname = "col1")
-#' mosub@matches
+#' ## Inspecting the matches in `mo`:
+#' mo$col1
+#' mo$target
 #'
-#' ## Manually adding matches using values of "col1" column in query and of 
-#' target
-#' moadd <- addMatches(mo, queryValues = q1[c(3, 5), "col1"],
-#'                     targetValues = t2[c(1, 6)], queryColname = "col1",
-#'                     scores = data.frame(score = c(1, 1.1)))
-#' #' ## Manually adding matches using indexes of query and target values to match
-#' moadd <- addMatches(mo, queryValues = c(3L, 5L), targetValues = c(1L, 6L),
-#'                     scores = data.frame(score = c(1, 1.1)), isIndex = TRUE)
+#' ## We have thus target *12* matched to both query elements with values 1 and
+#' ## 2, and query element 2 is matching 3 target elements. Let's assume we want
+#' ## to resolve this multiple mappings to keep from them only the match between
+#' ## query 1 (column `"col1"` containing value `1`) with target 1 (value `12`)
+#' ## and query 2 (column `"col1"` containing value `2`) with target 2 (value
+#' ## `13`). In addition we also want to keep query element 5 (value `5` in
+#' ## column `"col1"`) with the target with value `15`:
+#' mo_sub <- filterMatches(mo, queryValue = c(1, 2, 5), queryColname = "col1",
+#'     targetValue = c(12, 13, 15))
+#' matchedData(mo_sub)
+#'
+#' ## Alternatively to specifying the matches to filter with `queryValue` and
+#' ## `targetValue` it is also possible to specify directly the index of the
+#' ## match(es) in the `@matches` `data.frame`:
+#' mo@matches
+#'
+#' ## To keep only matches like in the example above we could use:
+#' mo_sub <- filterMatches(mo, index = c(1, 3, 5))
+#' matchedData(mo_sub)
+#'
+#' ## Note also that, instead of keeping the specified matches, it would be
+#' ## possible to remove them by setting `keep = FALSE`. Below we remove
+#' ## selected matches from the object:
+#' mo_sub <- filterMatches(mo, queryValue = c(2, 2), queryColname = "col1",
+#'     targetValue = c(12, 14), keep = FALSE)
+#' mo_sub$col1
+#' mo_sub$target
+#'
+#' ########
+#' ## Adding matches using `addMatches`
+#'
+#' ## `addMatches` allows to manually add matches. Below we add a new match
+#' ## between the `query` element with a value of `1` in column `"col1"` and
+#' ## the target element with a value of `15`. Parameter `score` allows to
+#' ## assign a score value to the match.
+#' mo_add <- addMatches(mo, queryValue = 1, queryColname = "col1",
+#'     targetValue = 15, score = data.frame(score = 1.40))
+#' matchedData(mo_add)
+#' ## Matches are always sorted by `query`, thus, the new match is listed as
+#' ## second match.
 #'
 #' ## Creating a MatchedSummarizedExperiment object.
 #' library(SummarizedExperiment)
@@ -685,12 +717,12 @@ pruneTarget <- function(object) {
     object
 }
 
-.findMatchesIdxs <- function(query, target, matches, queryValues = integer(), 
-                         targetValues = integer(), 
-                         queryColname = character(), 
+.findMatchesIdxs <- function(query, target, matches, queryValue = integer(),
+                         targetValue = integer(),
+                         queryColname = character(),
                          targetColname = character()) {
-  if(length(queryValues) != length(targetValues))
-    stop("'queryValues' and 'targetValues' must have the same length")
+  if(length(queryValue) != length(targetValue))
+    stop("'queryValue' and 'targetValue' must have the same length")
   if(length(dim(query)) == 2) {
     if(length(queryColname) == 0)
       stop("\"", queryColname,"\" must be set when 'query' is 2-dimensional")
@@ -704,12 +736,12 @@ pruneTarget <- function(object) {
     if(!targetColname %in% colnames(target))
       stop("\"", targetColname,"\" is not a column of 'target'")
   }
-  mq <- .extract_elements(query, matches$query_idx, queryColname, 
+  mq <- .extract_elements(query, matches$query_idx, queryColname,
                           drop = TRUE)
   mt <- .extract_elements(target, matches$target_idx, targetColname,
                           drop = TRUE)
-  unlist(sapply(seq_along(queryValues), function(i) 
-    which(mq == queryValues[i] & mt == targetValues[i])))
+  unlist(sapply(seq_along(queryValue), function(i)
+    which(mq == queryValue[i] & mt == targetValue[i])))
 }
 
 #' @rdname Matched
@@ -719,69 +751,70 @@ setGeneric("filterMatches", function(object, ...)
   standardGeneric("filterMatches"))
 
 #' @rdname Matched
-#' 
+#'
 #' @importFrom methods validObject
 #'
 #' @export
-setMethod("filterMatches", "Matched", 
-          function(object, queryValues = integer(), targetValues = integer(),
+setMethod("filterMatches", "Matched",
+          function(object, queryValue = integer(), targetValue = integer(),
                    queryColname = character(), targetColname = character(),
-                   idxs = integer(), keep = TRUE, ...) {
-            if(length(idxs) && any(!idxs %in% seq_len(nrow(object@matches))))
-              stop("some indexes in \"idxs\" are out of bounds")
-            if(!length(idxs) && length(queryValues))
-              idxs  <- .findMatchesIdxs(object@query, object@target, 
-                                        object@matches, queryValues, 
-                                        targetValues, queryColname, 
+                   index = integer(), keep = TRUE, ...) {
+            if(length(index) && any(!index %in% seq_len(nrow(object@matches))))
+              stop("some indexes in \"index\" are out of bounds")
+            if(!length(index) && length(queryValue))
+              index  <- .findMatchesIdxs(object@query, object@target,
+                                        object@matches, queryValue,
+                                        targetValue, queryColname,
                                         targetColname)
-            if(keep) to_keep <- seq_len(nrow(object@matches)) %in% idxs
-            else to_keep <- !seq_len(nrow(object@matches)) %in% idxs
+            if(keep) to_keep <- seq_len(nrow(object@matches)) %in% index
+            else to_keep <- !seq_len(nrow(object@matches)) %in% index
             object@matches <- object@matches[to_keep, , drop = FALSE]
             validObject(object)
             object
           })
 
-.addMatches <- function(query, target, matches, queryValues = integer(),
-                        targetValues = integer(), queryColname = character(),
-                        targetColname = character(), scores = data.frame(),
+.addMatches <- function(query, target, matches, queryValue = integer(),
+                        targetValue = integer(), queryColname = character(),
+                        targetColname = character(),
+                        score = data.frame(),
                         isIndex = FALSE) {
-  if(!is.data.frame(scores))
-    stop("'scores' must be a data.frame")
-  if(ncol(scores) != ncol(matches) - 2 || 
-     any(!colnames(scores) %in% colnames(matches)))
-    stop("columns of 'scores' incompatible with object matches")
-  if(length(queryValues) != length(targetValues) || 
-     length(queryValues) != nrow(scores))
-    stop("'queryValues', 'targetValues' and 'scores' must have the same length")
-  if(isIndex) {
-    if(!is.integer(queryValues) || !is.integer(targetValues))
-      stop(paste0("'queryValues' and 'targetValues' must be integer ", 
+  if (!is.data.frame(score))
+    stop("'score' must be a data.frame")
+  if (ncol(score) != ncol(matches) - 2 ||
+     any(!colnames(score) %in% colnames(matches)))
+    stop("columns of 'score' incompatible with object matches")
+  if (length(queryValue) != length(targetValue) ||
+      length(queryValue) != nrow(score))
+    stop("'queryValue', 'targetValue' and 'score' must have the same length")
+  if (isIndex) {
+    if (!is.integer(queryValue) || !is.integer(targetValue))
+      stop(paste0("'queryValue' and 'targetValue' must be integer ",
                   "vectors when 'isIndex == TRUE'"))
-    if(any(!queryValues %in% seq_len(.nelements(query))))
-      stop("Some of the provided indexes in 'queryValues' are out of bound")
-    if(any(!targetValues %in% seq_len(.nelements(target))))
-      stop("Some of the provided indexes in 'queryValues' are out of bound") 
-    query_idx <- queryValues
-    target_idx <- targetValues
+    if (any(!queryValue %in% seq_len(.nelements(query))))
+      stop("Some of the provided indexes in 'queryValue' are out of bound")
+    if (any(!targetValue %in% seq_len(.nelements(target))))
+      stop("Some of the provided indexes in 'queryValue' are out of bound")
+    query_idx <- queryValue
+    target_idx <- targetValue
   } else {
-    if(length(dim(query)) == 2)
-      if(!queryColname %in% colnames(query))
+    if (length(dim(query)) == 2)
+      if (!queryColname %in% colnames(query))
         stop("\"", queryColname,"\" is not a column of 'query'")
-    if(length(dim(target)) == 2)
-      if(!targetColname %in% colnames(target))
+    if (length(dim(target)) == 2)
+      if (!targetColname %in% colnames(target))
         stop("\"", targetColname,"\" is not a column of 'target'")
-    mq <- match(queryValues, .extract_elements(query, j = queryColname, 
+    mq <- match(queryValue, .extract_elements(query, j = queryColname,
                                                drop = TRUE))
-    mt <- match(targetValues, .extract_elements(target, j = targetColname, 
+    mt <- match(targetValue, .extract_elements(target, j = targetColname,
                                                 drop = TRUE))
     to_keep <- !is.na(mq) & !is.na(mt)
     query_idx <- mq[to_keep]
     target_idx <- mt[to_keep]
-    scores <- scores[to_keep, ]
+    score <- score[to_keep, ]
   }
   new_matches <- rbind(matches, data.frame(query_idx = query_idx,
                                            target_idx = target_idx,
-                                           score = scores))
+                                           score = score))
   # remove possible matches that were already in matches
   new_matches[!duplicated(new_matches[, c("query_idx", "target_idx")]), ]
 }
@@ -793,18 +826,18 @@ setGeneric("addMatches", function(object, ...)
   standardGeneric("addMatches"))
 
 #' @rdname Matched
-#' 
+#'
 #' @importFrom methods validObject
 #'
 #' @export
-setMethod("addMatches", "Matched", 
-          function(object, queryValues = integer(), targetValues = integer(),
-                   queryColname = character(), targetColname = character(), 
-                   scores = data.frame(), isIndex = FALSE) {
-            object@matches <- .addMatches(object@query, object@target, 
-                                          object@matches, queryValues, 
-                                          targetValues, queryColname, 
-                                          targetColname, scores, isIndex)
+setMethod("addMatches", "Matched",
+          function(object, queryValue = integer(), targetValue = integer(),
+                   queryColname = character(), targetColname = character(),
+                   score = data.frame(), isIndex = FALSE) {
+            object@matches <- .addMatches(object@query, object@target,
+                                          object@matches, queryValue,
+                                          targetValue, queryColname,
+                                          targetColname, score, isIndex)
             validObject(object)
             object
           })