+++ NEWS.md
@@ -1,5 +1,10 @@
 # MetaboAnnotation 0.2
 
+## Changes in 0.2.4
+
+- Support `data.frame`, `DataFrame` and `matrix` in `matchMz`.
+- Add `addMatches` and `filterMatches` functions.
+
 ## Changes in 0.2.3
 
 - Fixes in `MatchedSpectra`.